requirements.txt
annotated-types
anyio
boto3
botocore
cffi
click
colorama
cryptography
ecdsa
exceptiongroup
fastapi
greenlet
h11
idna
jmespath
mangum
pyasn1
pycparser
pydantic
pydantic-core
python-dateutil
python-dotenv
python-jose
PyYAML
rsa
s3transfer
six
sniffio
sqlalchemy
starlette
typing-extensions
typing-inspection
urllib3
uvicorn
pymysql
